Complete left side of page: banner, logo & credits
A bug was present that prevented the background image from extending
all the way down the page and limited to it to only the viewport. It was
fixed by removing a manual height of 100vh set on the right side's
container div which constrained the root element to only 100vh.